pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "filu-x"
version = "0.0.1"
description = "File-based decentralized social media"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "Mikael1979", email = "mikael@example.com"}
]

dependencies = [
    "click>=8.1",
    "cryptography>=41.0",
    "jinja2>=3.1",
    "python-slugify>=8.0",
    "requests>=2.31.0",
    "python-magic>=0.4.27",
    "bleach>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "mypy>=1.0",
    "isort>=5.0",
    "qrcode>=7.0",
]
ipfs = [
    "py-ipfs-http-client>=0.1.0",
]

[project.scripts]
filu-x = "filu_x.cli:cli"
